+++ front-end-tools/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased changes
 
+- Fix error message when specifying input parameter without uploading schema.
+- Fix error when contract calls return BigInts.
+- Fix contract method names in error messages.
+- Give warning when deploying a module that does not have embedded build information.
+
 ## 3.0.1
 
 - Upgrade dependencies @concordium/web-sdk and @concordium/react-components